HW-2 Add MovieSearch
